backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
openai==1.2.0
pymongo==4.6.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
transformers==4.35.0
torch==2.1.0
numpy==1.26.1
scikit-learn==1.3.2
python-multipart==0.0.6
aiohttp==3.8.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1